Renamed SimpleXmlWriter to SimpleXmlEncoder
The SimpleXmlWriter/Encoder does not write xml but just encodes it. Hence, the
class was renamed and moved from org.culturegraph.mf.stream.sink to
org.culturegraph.mf.stream.converter.xml.
