+++ src/main/resources/flux-commands.properties
@@ -71,7 +71,7 @@ object-tee org.culturegraph.mf.stream.pipe.ObjectTee
 stream-tee org.culturegraph.mf.stream.pipe.StreamTee
 wait-for-inputs	org.culturegraph.mf.stream.pipe.CloseSupressor
 
-stream-to-xml	org.culturegraph.mf.stream.sink.SimpleXmlWriter
+stream-to-xml	org.culturegraph.mf.stream.converter.xml.SimpleXmlEncoder
 rdf-macros	org.culturegraph.mf.stream.pipe.RdfMacroPipe
 
 batch-log	org.culturegraph.mf.stream.pipe.BatchLogger
